# solution/openapi.yaml
openapi: 3.0.3
info:
  title: PROD Backend 2025 Advertising Platform API
  version: "1.0.0"
  description: >
    API для управления данными клиентов, рекламодателей, рекламными кампаниями,
    показом объявлений, статистикой и управлением "текущим днём" в системе.
...
paths:
...
  /advertisers/{advertiserId}:
    get:
      tags:
        - Advertisers
      summary: Получение рекламодателя по ID
      description: Возвращает информацию о рекламодателе по его ID.
      operationId: getAdvertiserById
      parameters:
        - in: path
          name: advertiserId
          required: true
          description: UUID рекламодателя.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Информация о рекламодателе успешно получена.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Advertiser'
        '404':
          description: Рекламодатель не найден.
        '400':
          description: Некорректный идентификатор.
  /advertisers/bulk:
    post:
      tags:
        - Advertisers
      summary: Массовое создание/обновление рекламодателей
      description: Создаёт новых или обновляет существующих рекламодателей.
      operationId: upsertAdvertisers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/AdvertiserUpsert'
      responses:
        '201':
          description: Успешное создание/обновление рекламодателей.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Advertiser'
        '400':
          description: Переданы недействительные данные.
...
  /advertisers/{advertiserId}/generation/stream:
    post:
      tags:
        - Advertisers AI Features
      summary: Получение генерируемого текста для объявления с опорой на заголовок и название рекламодателя
      description: Возвращает текст для объявления сгенерированный искусственным интеллектом исходя из его названия и названия рекламодателя.
      operationId: getAITextStream
      parameters:
        - in: path
          name: advertiserId
          required: true
          description: UUID рекламодателя.
          schema:
            type: string
            format: uuid
      requestBody:
        description: Объект с одним полем - title, которое является заголовком желаемого объявления.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIGenerationRequest'
      responses:
        '200':
          description: Текст успешно сгенерирован.
          content:
            text/plain:
              schema:
                type: string
        '429':
          description: Превышен rate-limit ИИ.
        '404':
          description: Рекламодатель не найден.
        '400':
          description: Переданы недействительные данные.
  /advertisers/{advertiserId}/generation:
    post:
      tags:
        - Advertisers AI Features
      summary: Получение генерируемого текста для объявления с опорой на заголовок и название рекламодателя
      description: Возвращает текст для объявления сгенерированный искусственным интеллектом исходя из его названия и названия рекламодателя.
      operationId: getAIText
      parameters:
        - in: path
          name: advertiserId
          required: true
          description: UUID рекламодателя.
          schema:
            type: string
            format: uuid
      requestBody:
        description: Объект с одним полем - title, которое является заголовком желаемого объявления.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIGenerationRequest'
      responses:
        '200':
          description: Текст успешно сгенерирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIResponse'
        '429':
          description: Превышен rate-limit ИИ.
        '404':
          description: Рекламодатель не найден.
        '400':
          description: Переданы недействительные данные.
  # Рекламные кампании
...
  /advertisers/{advertiserId}/campaigns/{campaignId}/image:
    post:
      tags:
        - Campaigns Images
      summary: Устанавливает изображение рекламной кампании
      description: Устанавливает картинку рекламной кампании рекламодателя по заданному campaignId.
      operationId: setCampaignImage
      parameters:
        - in: path
          name: advertiserId
          required: true
          description: UUID рекламодателя, которому принадлежит кампания.
          schema:
            type: string
            format: uuid
        - in: path
          name: campaignId
          required: true
          description: UUID рекламной кампании, которой нужно установить изображение.
          schema:
            type: string
            format: uuid
      requestBody:
        description: Картинка рекламной кампании.
        required: true
        content:
          image/png:
            schema:
              type: string
              format: binary
          image/jpeg:
            schema:
              type: string
              format: binary
      responses:
        '201':
          description: Картинка рекламной кампании успешно установлена.
        '404':
          description: Рекламодатель или кампания не найдены.
        '400':
          description: Переданы недействительные данные.
    delete:
      tags:
        - Campaigns Images
      summary: Удаление картинки рекламной кампании
      description: Удаляет картинку с рекламной кампании рекламодателя по заданному campaignId.
      operationId: deleteCampaignImage
      parameters:
        - in: path
          name: advertiserId
          required: true
          description: UUID рекламодателя, которому принадлежит кампания.
          schema:
            type: string
            format: uuid
        - in: path
          name: campaignId
          required: true
          description: UUID рекламной кампании, картинку которого необходимо удалить.
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Картинка рекламной кампании успешно удалена.
        '404':
          description: Рекламодатель или кампания не найдены.
        '400':
          description: Переданы недействительные данные.
  # Рекламные объявления и клики
...
  /ads/{adId}/image:
    get:
      tags:
        - Ads
      summary: Получение рекламного изображения
      description: Возвращает изображение для рекламы (если оно имеется)
      operationId: getAdImageForClient
      parameters:
        - in: path
          name: adId
          required: true
          description: UUID рекламного объявления (идентификатор кампании), по которому необходимо изображение.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Картинка рекламной кампании передана пользователю.
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
        '404':
          description: Картинка не найдена.
        '400':
          description: Переданы недействительные данные.
  # Статистика
...
components:
  schemas:
    AIResponse:
      type: object
      properties:
        text:
          type: string
      required:
        - text
    AIGenerationRequest:
      type: object
      properties:
        title:
          type: string
      required:
        - title
...
    Advertiser:
      type: object
      description: Объект, представляющий рекламодателя.
      properties:
        advertiser_id:
          type: string
          format: uuid
          description: Уникальный идентификатор рекламодателя (UUID).
        name:
          type: string
          description: Название рекламодателя.
      required:
        - advertiser_id
        - name
    # --- ML скор ---
...
    AdvertiserUpsert:
      type: object
      properties:
        advertiser_id:
          type: string
          format: uuid
          description: UUID рекламодателя.
        name:
          type: string
      required: [ advertiser_id, name ]